Add function to rotate the root keys
This adds a function to rotate the root keys stored in the bootloader
by overwriting them with new keys.

This is intended to be an intermediate firmware release, dedicated to
rotating the keys. Only official firmware releases are able to rotate
the root keys.

When booted, it will rotate the keys and then reboot into the
bootloader, so that regular firmware can be installed.